# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1.0/roles": {
            "get": {
                "summary": "Show roles",
                "description": "Returns a list of all roles with pagination data.",
                "operationId": "add43a2369a12503d32ce0ab6f078b29",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number to retrieve.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "The number of results per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of roles.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pagination": {
                                            "properties": {
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                },
                                                "total_items": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Role"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new role",
                "description": "Creates a new role in the system.",
                "operationId": "7a42de25c319df0d0006f151f3b4def4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role",
                                    "description"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1.0/roles/{id}": {
            "get": {
                "summary": "Show a specific role",
                "description": "Fetches the details of a single role.",
                "operationId": "14b6097328b4aa180f564be349a583a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the role to retrieve.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the role details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update an existing role",
                "description": "Updates an existing role in the system.",
                "operationId": "9e0144fcebe39d42a74630e6ec5978ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the role to update.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated the role.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a specific role",
                "description": "Deletes a role from the system.",
                "operationId": "4bdce29db3838c6d0a749b5c4499106d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the role to delete.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role successfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Role": {
                "required": [
                    "role",
                    "description"
                ],
                "properties": {
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Administrator role"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}